gives user feedback when comment has / has not been posted sucessfully

+++ src/components/Comments/DeleteComment.jsx
@@ -1,20 +1,31 @@
 import { deleteComment } from "../../assets/api-calls";
 import Button from "react-bootstrap/Button";
+import { UserContext } from "../Users/UserContext";
+import { useContext } from "react";
+import { useState } from "react";
 
-function DeleteComment({ comment_id, onDelete }) {
+function DeleteComment(props) {
+  const { comment_id, onDelete, comment_author } = props;
+  const { user } = useContext(UserContext);
+  const [loading, isLoading] = useState(true);
+  const [err, setError] = useState(false);
+
   const handleDelete = () => {
-    deleteComment(comment_id).then((response) => {
-      console.log(response);
-      onDelete(comment_id);
-    });
+    if (user === comment_author) {
+      deleteComment(comment_id).then((response) => {
+        isLoading(true);
+        onDelete(comment_id);
+        isLoading(false);
+      });
+    }
   };
 
+  if (loading) {
+    return <p className='loading'>Your comment is being loaded...</p>;
+  }
+
   return (
-    <Button
-      variant='primary'
-      size='lg'
-      onClick={() => handleDelete()}
-      type='submit'>
+    <Button variant='primary' size='lg' onClick={handleDelete} type='submit'>
       {" "}
       Delete comment
     </Button>
